+++ app/src/main/AndroidManifest.xml
@@ -28,7 +28,7 @@
             android:label="@string/app_name"
             />
         <activity
-            android:name=".activites.MainActivity"
+            android:name=".activites.identity.IdentityManagementActivity"
             android:label="@string/title_activity_advanced"
             />
