+++ app/src/main/java/org/ea/sqrl/activites/create/NewIdentityDoneActivity.java
@@ -11,7 +11,6 @@ import android.widget.TextView;
 import org.ea.sqrl.R;
 import org.ea.sqrl.activites.SimplifiedActivity;
 import org.ea.sqrl.activites.identity.ExportOptionsActivity;
-import org.ea.sqrl.activites.MainActivity;
 import org.ea.sqrl.activites.base.LoginBaseActivity;
 import org.ea.sqrl.processors.SQRLStorage;
