+++ app/src/main/res/layout/activity_identity_management.xml
@@ -4,8 +4,8 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:id="@+id/mainActivityView"
-    tools:context="org.ea.sqrl.activites.MainActivity">
+    android:id="@+id/identityManagementActivityView"
+    tools:context="org.ea.sqrl.activites.identity.IdentityManagementActivity">
 
     <ImageButton
         android:id="@+id/btnCloseMain"
